Create redux
Instructions

Below is a pared-down demonstration of React and Redux working together. It has the essential elements you need to connect/subscribe a React component to the Redux store for data, to dispatch actions, and to update the store's state using a reducer.

Challenge 1: Fix the existing code

When you enter a name and a course and hit the "Add Instructor" button, it should update the instructors list in the view. However, it isn't working right now, and lucky you, you get to debug it. Note: This is more of a React review challenge than a Redux one.

    Why is the instructorsList component not rendering? Fix this first so that you can see the list. Hint: It's a small but important typographical error.
    In the AddInstructor component, the event handler has not been bound correctly. Can you fix it? There are two different and acceptable ways to do this. Refer to this demo -- specifically, GoodExample1 and ArrowExample2.

Challenge 2: Update the fixed code

For this part of the challenge, add a 'Color' field to the form, next to "Course Taught". Then update the code so that the background color of the instructor name renders as the color that you enter. You will need to update multiple places in the code for this to work: in the reducer, the action, and a couple of other places in addition to the form.

Note: In the reducer, instructor.color is currently hardcoded to black (#000). When you are at this part of the challenge, remove this line completely along with the other changes.

Good luck!